Update to integrity test
