+++ public/views/_integrity.html
@@ -311,11 +311,11 @@
     <script type="module">
         import { runWebTests } from "codi";
 
-        const testFiles = [`${window.location.origin}{{dir}}/public/tests/_default.test.mjs`];
+        const testFiles = [`${window.location.origin}{{dir}}/public/tests/_integrity.test.mjs`];
 
         try {
             let results = await runWebTests(testFiles);
-            
+
             const urlParams = new URLSearchParams(window.location.search);
             const shouldDownload = urlParams.get('download') === 'true';
